Performance script tested in jupyter, deleted deplicate files
